GH-323: [refactor] Group together event related file under the same package
