+++ Microservices/event-service/src/main/java/app/sportahub/eventservice/dto/request/event/ParticipantRequest.java
@@ -1,4 +1,4 @@
-package app.sportahub.eventservice.dto.request;
+package app.sportahub.eventservice.dto.request.event;
 
 import app.sportahub.eventservice.model.event.participant.ParticipantAttendStatus;
 import com.fasterxml.jackson.annotation.JsonInclude;
