+++ Microservices/event-service/src/main/java/app/sportahub/eventservice/dto/request/event/TeamRequest.java
@@ -1,4 +1,4 @@
-package app.sportahub.eventservice.dto.request;
+package app.sportahub.eventservice.dto.request.event;
 
 import com.fasterxml.jackson.annotation.JsonInclude;
 import jakarta.validation.constraints.NotBlank;
