+++ Microservices/event-service/src/main/java/app/sportahub/eventservice/repository/event/EventRepository.java
@@ -1,8 +1,8 @@
-package app.sportahub.eventservice.repository;
+package app.sportahub.eventservice.repository.event;
 
 import app.sportahub.eventservice.model.event.Event;
-import org.springframework.data.domain.Pageable;
 import org.springframework.data.domain.Page;
+import org.springframework.data.domain.Pageable;
 import org.springframework.data.mongodb.repository.MongoRepository;
 import org.springframework.stereotype.Repository;
 
@@ -12,7 +12,10 @@ import java.util.Optional;
 public interface EventRepository extends MongoRepository<Event, String> {
 
     Optional<Event> findEventById(String id);
+
     Optional<Event> findEventByEventName(String eventName);
+
     Page<Event> findByParticipantsUserId(String userId, Pageable pageable);
+
     Page<Event> findByCreatedBy(String userId, Pageable pageable);
 }
